Fix obstacle avoidance module
Storing some intermediate data in an union to save some space is a good
idea but makes the code not very portable. As we have a lot more RAM now
than in 2004, we can safely move it out of an union.

Also add some basic smoke tests for the obstacle avoidance module,
showcasing the API.