Implement bus watchgroups
Bus watchgroups are my solution to the "select" problem: how can we wait
on many topics in a single thread in a portable way? This is required,
e.g. to forward topics over UDP, log them, or whatever.

The chosen approach is to group topics under a single condition variable
that will be signaled every time a topic in the group is published to.

This commit also adds a demo application showcasing this feature by
implementing an observer monitoring two topics.