Allow topics to be added to several groups
In order to do so each topic is given a linked list of groups to which
it is subscribed, instead of directly linking groups together. This
indirection makes the API a bit less clean because of the added
watchers but is required in order to be able to add a thread to two
groups (see additional tests).

Thanks to @Stapelzeiger for catching this.