Remove the EnableMotor service call
This service call was used to enable or disable the power output of the
motor boards. I decided to remove it for two reasons:

1. It did not provide any additional features: if no setpoints are sent
in a timely manner (300 ms currently) the boards will disable
themselves.
2. It makes the init sequence more complicated than needed: now you only
have to send the config before being ready to send setpoints.

I also have a long term plan to remove as much custom DSDL from our
repository as possible, and this is a step in the right direction.