Manually manage reentrancy for printf
It appears that most Newlib functions are not thread safe, and this may
cause problems in some cases. To prevent this, we can use the reentrant
variants manually.

Another option is to recompile Newlib with the __DYNAMIC_REENT__ flag
defined and implement the __getreent syscalls. However it is not really
convenient to ask the whole team to build Newlib locally.