Integrate streams into the parameter tree
Previously this was a part of the board that was not configurable
through the parameter tree. This means the configuration would not be
saved into the flash and reloaded on boot.

This commit does not remove the legacy way of setting up streams, but
only adds the standard, parameter-based method. Once this is properly
integrated, the legacy setup can be removed (will also affect
master-firmware).