Tweak cube deposit sequence when build tower
The window is increased to avoid hanging on given setpoints
I don't know why this is happening only here, but that's a
fix that worked for me.

Also tuned the arm trajectory waypoints to make it correct.
i.e. avoid blocking when putting cube on table by pressing
and never being able to reach goal