Explicit ownership of obstacle_avoidance
Previously, obstacle_avoidance used to be a global singleton, leading to
all sorts of issues with shared ownership, proper locking and so on.

This commit changes the API so that all calls to the obstacle avoidance
module take a pointer to the data to use. This also meant changing a few
tests.

On the robot the obstacle avoidance data is now handled by the map
server. The only way to access it is through an API that ensures it is
properly locked.

Fixes #137