Motor: Expose current offset to be calibrated
We've observed that there is a constant current offset
measured by the motor board on all motor boards with a
Molex picoblade connected on them (5 to 10mA), although
the circuit is not closed so no current should flow.
The goal here is provide a way to compesante for this
flaw, so we can make the current PID loop more stable.
Otherwise we can't control the proximity beacon's
current and velocity properly