clang-tidy: Mark overriden functions explicitely.
From the Google C++ style guide[0]: "Explicitly annotate overrides of
virtual functions or virtual destructors with exactly one of an override
or (less frequently) final specifier. Do not use virtual when declaring
an override. Rationale: A function or destructor marked override or
final that is not an override of a base class virtual function will not
compile, and this helps catch common errors. The specifiers serve as
documentation; if no specifier is present, the reader has to check all
ancestors of the class in question to determine if the function or
destructor is virtual or not."

[0] https://google.github.io/styleguide/cppguide.html#Inheritance