Simulate a movable opponent
This commit adds a simulated opponent that can be moved around by left
clicking on the simulated table. It is cleanly separated between a
physics object (a static body) and the viewer. The goal is to be able to
add the UAVCAN beacon publisher, so that we are able to test obstacle
avoidance strategies.