Synchronize with external github repository
This adds a copybara[0] script to export changes to our own separate C++
repository [1] for ease of use by other persons. This allows us to
maintain the source of truth inside the CVRA repository, but still let
other people use our implementation.

It includes a license,as well as an example build.

[0] https://github.com/google/copybara
[1] https://github.com/cvra/goap-cpp