Convert opponent position in the beacon handler
Converting this here makes more sense as the conversion from a range and
bearing to an absolute position is a concern specific to the type of
beacon used, and not a general thing. It allows other modules in our
code to not care about how the conversion is done. This in turn makes
for example unit testing easier.