trace: Add support for data points without arguments
Until then the data points without arguments were just data point with
empty string. However, this made the trace output considerably harder to
read. The tradeoff this creates is that in the same memory we can now
run only 127 different trace points instead of 256, but I think this is
an acceptable tradeoff for now.