Add support for per-board configurations
This implements a mechanism to define per-board configuration overloads.
They are applied over the default parameter value, but any stored
configuration is still applied over them. This allows us to centrally
define the configuration for a bunch of boards, which comes in handy
when flashing many of them, for example by DFU.

This commit also contains support code for this feature, namely a way to
easily find the processor's UID from outside (either over USB, or over
CAN). It also creates a configuration file for those new parameters.