Implement rate limiting to protect the CAN bus
The CAN bus can easily be overloaded if we have a lot of data coming in.
Since we are actually buffering in the Python script, that could lead to
unbounded buffer, which are terrible for IP performance, in particular
with TCP. Instead we should do what proper routers do, and drop the
packet.