Proof of concept of automatic simulation recording
The long term idea behind this is to be able to record simulated matches
as part of the CI process, upload the video and make it easily
accessible. This would allow everyone to check the result of a match.

This commit is the first part, where we extract the video from webot. We
now need to build a small web interface that allows you to access those
recordings.